Add related-record & related-tag for Record(Attention:did not test);set project include & link path with MSVC's micro; delete *.user files;
Former-commit-id: f6cb8a63a27257b19dca157f8e66094ae06cf5bb [formerly 861a9538c009765616b64dcd1a1076f8de1a767a]
Former-commit-id: ed7ce3fde7bb4af503913f533f87ce49479ed9bc

+++ NFComm/Config/NFConfig.h
@@ -31,7 +31,7 @@ public:
 
     const int Count()
     {
-        return map.size();
+        return static_cast<int>(map.size());
     }
 
     // The actual map used to store key/value pairs
